feat: report functions implementation
