+++ consertos.json
@@ -1,1 +1,18 @@
-{}
+{
+    "1_ABC_10/06/2024": {
+        "CPF": "1",
+        "Placa": "ABC",
+        "Data de Entrada": "10/06/2024",
+        "Data de Sa\u00edda": "11/06/2024",
+        "Descri\u00e7\u00e3o dos Problemas": "Muito ruim",
+        "Valor de Conserto": 200.0
+    },
+    "2_DEF_11/06/2024": {
+        "CPF": "2",
+        "Placa": "DEF",
+        "Data de Entrada": "11/06/2024",
+        "Data de Sa\u00edda": "12/06/2024",
+        "Descri\u00e7\u00e3o dos Problemas": "Muito p\u00e9ssimo",
+        "Valor de Conserto": 300.0
+    }
+}
